Extracted the Profile component logic into a new useProfile custom hook
