+++ src/features/Profile/index.js
@@ -1,37 +1,22 @@
-import { useEffect } from "react";
-import { useParams } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import {
-    getPersonId,
-    selectDetailsStatus,
-    selectPeopleDetails,
-    selectGenresDetails,
-    selectPeopleCast,
-    selectPeopleCrew,
-    fetchPeopleDetailsLoad
-} from "./peopleDetailsSlice";
+import { useProfile } from "./useProfile";
 import PersonInfo from "./PersonInfo";
 import PersonMovies from "./PersonMovies";
 import Loading from "../Actions/Loading";
 import Error from "../Actions/Error";
 
 const Profile = () => {
-    const dispatch = useDispatch();
-    const { personId } = useParams();
-
-    const status = useSelector(selectDetailsStatus);
-    const personDetails = useSelector(selectPeopleDetails);
-    const movieGenres = useSelector(selectGenresDetails);
-    const personCast = useSelector(selectPeopleCast);
-    const personCrew = useSelector(selectPeopleCrew);
-
-    useEffect(() => {
-        dispatch(fetchPeopleDetailsLoad(personId));
-    }, [dispatch, personId]);
-
-    useEffect(() => {
-        dispatch(getPersonId((personId)));
-    }, [dispatch, status, personId]);
+    const {
+        status,
+        movieGenres,
+        personCast,
+        personCrew,
+        personName,
+        personPhotoPath,
+        personBirthday,
+        personPlaceOfBirth,
+        personBiography,
+        formatDate
+    } = useProfile();
 
     if (status === "error") {
         return <Error />;
@@ -39,7 +24,14 @@ const Profile = () => {
 
     return status === "success" ? (
         <>
-            <PersonInfo personDetails={personDetails} />
+            <PersonInfo
+                personPhotoPath={personPhotoPath}
+                personName={personName}
+                personBirthday={personBirthday}
+                formatDate={formatDate}
+                personPlaceOfBirth={personPlaceOfBirth}
+                personBiography={personBiography}
+            />
             {personCast.length > 0 && (
                 <PersonMovies
                     title="cast"
